move search code into search.py
