+++ src/sourmash/search.py
@@ -1,12 +1,144 @@
+"""
+Code for searching collections of signatures.
+"""
 from collections import namedtuple
 import sys
+import os
+from enum import Enum
 
 from .logging import notify, error
 from .signature import SourmashSignature
 from .minhash import _get_max_hash_for_scaled
 
 
-# generic SearchResult.
+class SearchType(Enum):
+    JACCARD = 1
+    CONTAINMENT = 2
+    MAX_CONTAINMENT = 3
+
+
+def make_jaccard_search_query(do_containment, do_max_containment, best_only,
+                              threshold):
+    """\
+    Make a "flat" search object for Jaccard search & containment.
+    """
+    if do_containment and do_max_containment:
+        raise TypeError("'do_containment' and 'do_max_containment' cannot both be True")
+
+    # configure search - containment? ignore abundance? best only?
+    search_cls = JaccardSearch
+    if best_only:
+        search_cls = JaccardSearchBestOnly
+
+    if do_containment:
+        search_obj = search_cls(SearchType.CONTAINMENT, threshold)
+    elif do_max_containment:
+        search_obj = search_cls(SearchType.MAX_CONTAINMENT, threshold)
+    else:
+        search_obj = search_cls(SearchType.JACCARD, threshold)
+
+    return search_obj
+
+
+def make_gather_query(query_mh, threshold_bp):
+    "Make a search object for gather."
+    scaled = query_mh.scaled
+    if not scaled:
+        raise TypeError("query signature must be calculated with scaled")
+
+    # are we setting a threshold?
+    threshold = 0
+    if threshold_bp:
+        # if we have a threshold_bp of N, then that amounts to N/scaled
+        # hashes:
+        n_threshold_hashes = threshold_bp / scaled
+
+        # that then requires the following containment:
+        threshold = n_threshold_hashes / len(query_mh)
+
+        # is it too high to ever match? if so, exit.
+        if threshold > 1.0:
+            return None
+
+    search_obj = JaccardSearch(SearchType.CONTAINMENT, threshold=threshold)
+
+    return search_obj
+
+
+class JaccardSearch:
+    """
+    A class used by Index classes for searching/gathering.
+    """
+    def __init__(self, search_type, threshold=None):
+        "Constructor. Takes type of search, and optional threshold."
+        score_fn = None
+        require_scaled = False
+
+        if search_type == SearchType.JACCARD:
+            score_fn = self.score_jaccard
+        elif search_type == SearchType.CONTAINMENT:
+            score_fn = self.score_containment
+            require_scaled = True
+        elif search_type == SearchType.MAX_CONTAINMENT:
+            score_fn = self.score_max_containment
+            require_scaled = True
+        self.score_fn = score_fn
+        self.require_scaled = require_scaled # @CTB
+
+        if threshold is None:
+            threshold = 0
+        self.threshold = float(threshold)
+
+    def check_is_compatible(self, sig):
+        """
+        Is this query compatible with this type of search? Raise TypeError
+        if not.
+        """
+        if self.require_scaled:
+            if not sig.minhash.scaled:
+                raise TypeError("this search requires a scaled signature")
+
+        if sig.minhash.track_abundance:
+            raise TypeError("this search cannot be done with an abund signature")
+
+    def passes(self, score):
+        "Return True if this score meets or exceeds the threshold."
+        if score and score >= self.threshold:
+            return True
+        return False
+
+    def collect(self, score):
+        "Is this a potential match?"
+        pass
+
+    def score_jaccard(self, query_size, shared_size, subject_size, total_size):
+        "Calculate Jaccard similarity."
+        return shared_size / total_size
+
+    def score_containment(self, query_size, shared_size, subject_size,
+                          total_size):
+        "Calculate Jaccard containment."
+        if query_size == 0:
+            return 0
+        return shared_size / query_size
+
+    def score_max_containment(self, query_size, shared_size, subject_size,
+                              total_size):
+        "Calculate Jaccard max containment."
+        min_denom = min(query_size, subject_size)
+        if min_denom == 0:
+            return 0
+        return shared_size / min_denom
+
+
+class JaccardSearchBestOnly(JaccardSearch):
+    "A subclass of JaccardSearch that implements best-only."
+    def collect(self, score):
+        "Raise the threshold to the best match found so far."
+        self.threshold = max(self.threshold, score)
+
+
+# generic SearchResult tuple.
 SearchResult = namedtuple('SearchResult',
                           'similarity, match, md5, filename, name, query, query_filename, query_name, query_md5')
