+++ tests/test_index.py
@@ -10,7 +10,7 @@ import shutil
 import sourmash
 from sourmash import load_one_signature, SourmashSignature
 from sourmash.index import (LinearIndex, MultiIndex, ZipFileLinearIndex,
-                            get_search_obj)
+                            make_jaccard_search_query)
 from sourmash.sbt import SBT, GraphFactory, Leaf
 from sourmash.sbtmh import SigLeaf
 from sourmash import sourmash_args
@@ -64,7 +64,7 @@ def test_simple_index(n_children):
     linear.insert(leaf4_sig)
     linear.insert(leaf5_sig)
 
-    search_fn = get_search_obj(True, False, False, 0.0)
+    search_fn = make_jaccard_search_query(True, False, False, 0.0)
 
     kmers = ["AAAAA", "AAAAT", "AAAAG", "CAAAA", "GAAAA"]
     for kmer in kmers:
